+++ graphistry/compute/__init__.py
@@ -1,5 +1,4 @@
 from .ComputeMixin import ComputeMixin
 from .ast import (
-    n, e_forward, e_reverse, e_undirected
+    n, e_forward, e_reverse, e_undirected, is_in, IsIn
 )
-from .filter_by_dict import is_in, IsIn